requirements.txt
crewai==0.22.5
crewai-tools
python-dotenv
streamlit